jesd204: add state transitions for unregistering a device
This should de-initialize all JESD204 links in a certain order:
1. disable links
2. disable clocks
3. uninitialize devices

This still needs some more testing to make sure that it actually works ok.

The states should transition as defined, and the callbacks should be called
as defined (because they rely on the core state transition mechanism).

But who knows what else is incomplete?

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>